django_admin/requirements.txt
asgiref==3.7.2
bitrix24-rest==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
idna==3.6
Pillow==10.1.0
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.15.3
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
urllib3==2.2.0
tbank_kassa_api==0.0.2.7.6
